Refactored some of the tests so I can ensure they run in an order which helps to troubleshoot a problem.  In general an earlier test will fail if it tests a feature which a later test might rely on.  This makes writing later tests a little bit easier as you don't have to check for as much and if they fail you are pretty sure the relevant earlier tests also failed.
Added significantly more testing for the Tree object and
how it tracks its modified flag in memory.

Implemented creating of a repository from within Java.
This repository should match what a git-init-db would do
and makes it easier for us to construct JUnit test cases
as we can build (and destroy) repositories on the fly for
each test similiar to what the core git utilities do.

Patch: Prototype-Implementation