Finished implementing delete/rename hook.
The hook attempts to update the cache tree when moves
or deletes occur in the workspace.  Right now it does a
pretty good job at tracking the rename or delete but it
doesn't ensure that if the rename is going through that it
actually is going through as the cache may not get written
to disk before a workspace crash.

Patch: Prototype-Implementation