Added support for the new loose object format.
jgit now writes the new loose object format by default, but can
revert to the old header format if core.legacyHeader is set in the
config file.  When reading a loose object jgit will autodetect its
format and handle either.

The new loose object format allows a pack creater (such as the
command line git-pack-objects tool) to reuse the compressed data in
the loose object rather than having to decompress it and recompress
it, as the ASCII object header is no longer part of the compressed
data.

This change updated the unit tests as the new loose object format
was slightly smaller for an empty tree.  Additional unit tests are
still needed to verify both code paths work properly.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>