Implemented support for reading delta offsets in pack files.
Core git v1.4.4-rc1 and later contains support for creating pack files
which use an absolute offset from the start of the pack to denote the
base object for a delta, rather than the larger 20 byte SHA1 id.  This
reduces most pack files by 3.2-5%.

Commit eb32d236df0c16b936b04f0c5402addb61cdb311 by Nicolas Pitre over
in the core git repository provides a good explanation of this change.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>