Moved unit jgit tests.
Due to the refactoring applied during the last commit T0004_PackReader.java
can't access some of the APIs it was requesting, especially since it was
in a the lib_test package.  So I'm moving the tests all into the same
package as the library itself to permit access to some of these non-public
APIs.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>