Reformatted all source code to the Java Conventions that ship with Eclipse.
This is at the request of Robin Rosenberg who wisely pointed out that
the prior convention causes problems with 3 line context diffs as most
of the context will be { and } and almost nothing else, allowing the
patch to apply at almost any location within a source file.

This is a horribly ugly change that makes patch application based on
older versions impossible, but better we do it now while this project
is still very young then do it later when we're getting a patch a day.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>