Always check for a packed object before a loose object.
Make jgit do what core Git does, which is look for an object in all known
packs before going to the loose objects directory.

This is really a good idea as pack access tends to be very cheap compared
to the existing object directory, especially since the majority of the
objects tend to live only in pack files and only the most recent handful
of objects are loose.  Consequently most of the time the needed object
will be found in a pack.

This provides a decent performance boost when accessing lots of objects,
such as walking a history chain for a branch. 

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>