Refactor how MapMode gets selected.
Apparently we aren't allowed to access a read/write mmap as a byte[] in
Java.  I originally thought you could, but after watching it in the
debugger and reading the source code it looks like that isn't actually
allowed.  So there's no point in letting the caller select the MapMode
for a file; we either want no mmap (use read and byte[]s) or read only
mmap.  This is better represented by a simple boolean value.

Apparently the MappedByteBuffer overhead is also very high.  If we
access the same commit chain several times the relatively high cost
of initially reading in our windows gets ammortized out somewhat and
its faster to just use read+byte[]s:

On the Mozilla repository, walking a chain of 50,000 commits 5 times
using the same WindowCache gives us:

  usemap: 6373 ms
  nommap: 4677 ms

This is a pretty significant difference, almost 1.6 seconds.  Of course
core Git can perform the same traversal in just 1.8 seconds so jgit still
has plenty of performance problems.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>