Basic Git history support framework for Eclipse.
This adds history support classes for Eclipse (3.2 and newer), i.e.
most of the classes implied by the IFileHistoryProvider.

The IFileHistoryProvider.SINGLE_REVISION is used for QuickDiff
which means it points at the revision you want to see when editing.
For plain Git is the revision in HEAD, but for StGit it is the
revision the current patch is based on, i.e. HEAD-1, unless the stack is
empty, in which case it is HEAD.

This is not all functionality required for a full plugin, just enough
for quickdiff, simple history browsing and compare-with functionality.

Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>