Indroduce GitAction class
I introduce GitAction as base class for all actions. It allows to
disable the action menu item if the selection does not live in a
Git controlled repository.

The current code is quite "gruik". It is mainly based on copy/paste
from other plugins.

A better solution could be to avoid inserting EGit actions in menus
if the selection does not live in a Git repo. (some plugin.xml hack)

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>