Reformatted according to Eclipse style settings.
Because its far easier to maintain Java code where Eclipse handles all
of the formatting and layout, I'm forcing every chunk of Java code in
this project to the already agreed upon Eclipse style settings.  A
few items apparently have been added by myself and others since the
last bulk formatting and these apparently do not conform to the style.

In the future I'm probably going to start rejecting any changes which
do not conform to the style, as its more difficult to generate patches,
apply patches, or to just hand-edit code when you cannot rely on the
standard Eclipse formatter to do the heavy lifting for you.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>