Created notes on how to properly submit patches.
This is a blatent borrowing of the Documentation/SubmittingPatches
file from git.git.  I like Junio's development practices and how
he manages git, and how patches get submitted is a part of that
process.  There are however some minor differences with egit/jgit
as we are relying heavily on the Eclipse code formatter rather than
doing that sort of annoying work by hand.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>