Only checkpoint the updates, not the entire workspace.
The Checkpoint Project action was added by Robin to avoid the very
expensive background process which was firing off every few minutes
to update the cache tree.  The intent of the action was to allow the
user to force the checkpoint when they are ready to let Eclipse waste
time doing something, rather than having Eclipse steal their system at
random times.

Unfortunately a fullUpdate() causes the entire cache tree to be reset
to whatever is currently in the workspace.  This causes user changes
such as file additions or removals to not be reflected properly, and
it also just takes very long as we must recompute the full SHA1 of
every file within the project.

Instead we should just tie the action to checkpointIfNecessary(),
which is what the background job was invoking on a periodic basis.
This method only updates the parts of the cache tree which are dirty,
and it maintains the user's modifications which have not yet been
written to disk.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>