Move quickdiff code into the UI plugin.
Robin originally added the quickdiff support in the core plugin, probably
because he needed to access the RepositoryProvider class and found the
method was available from GitProvider.  Its actually just inherited from
the base RepositoryProvider supplied by Eclipse.

So I'm moving the quickdiff support to the UI plugin, as it is strictly
a UI feature.  This restores the dependency set for the core plugin to
only core workbench functionality, thereby making the plugin still
useful in a non-UI setting.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>