Make use of trace, logError in GitQuickDiffProvider.
Our UI plugin has a handy trace feature builtin to its Activator class.
We use this throughout most of the other code, so we should make use
of it in the quickdiff provider too.

I've also converted the stack trace printing to log the error with
the Eclipse Error Log, so that users can view the trace easily from
within the workbench (say to report a bug).

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>