The Great Reformat, Part II.
Robin's recent settings change modified the formatting standards to use
tabs rather than spaces for leading indentation.  I agree with those
(otherwise I wouldn't have applied the patch) but our source did not
conform as earlier we were using spaces for indentation.  The earlier
usage of spaces was my own fault and was not intentional.

This is a change of nearly every line, to make every file conform to
the project-specific settings included with EGit/JGit.  I cannot
stand not being able to reformat the entire file from within the Java
editor and have the entire file change, because none of it conforms to
the settings.

There are absolutely *no* logic changes in this commit.  If you are
bug hunting and bisect onto here, its the same as its parent.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>