Modified formatting style from 8-tab to 4-tab.
8 spaces per tab in Java is horrible.  Every single line of code is
starting at column 16 at best; toss in a single if statement and
you are already at column 24.  Wrap that if around a loop and what
was once a reasonable chunk of code in C (starting at column 8)
is now all the way in at column 32.  When you then tack in the
relatively long type names and method names that are common in
Java, almost no statement fits on a single line anymore.

One might be able to argue that an 8-spaces-per-tab intendation style
encourages the developer to make smaller functions when nesting gets
too deep, but in Java losing 8 columns to the class wrapper itself
just makes a mess of things, as you would then tend to start to abstract
single line statements into methods just to keep simple lines from
wrapping.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>