Resolve race condition with decorator thread during connect.
This change actually fixes a few issues.  The most noticable issue is
the race condition between the background thread which is creating the
initial cache tree during a connect operation, and the UI decorator
thread.  If the project is of any size the connect operation will take
a while and Eclipse will let the UI decorator thread to update the UI
before the background thread finishes creating the cache tree state.

The problem with that is the decorator only updates once, so it marks
files the background job hasn't gotten to yet as modified.  Once marked
as modified the decorator never received updates to tell it the files
aren't really modified.  This left the user with an inconsistent view
of their workspace.  Closing and reopening the project worked only
because it forced the decorator to reapply all formatting.

The problem also happened with the resource change listener.  Our GitData
class was monitoring the workspace for relevant resource change events
and updating the cache tree in-memory to be dirty, but it was not telling
the decorator that a change happened.  As a result the decorator was way
behind and did not reflect user actions as quickly as it was supposed to.

Now we notify the decorator anytime we remerge the cache tree with HEAD,
as this can have a possible change in resource status.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>