Ignore parameter assignment errors.
I've put some thought into this one, and I really dislike the parameter
assignment warning in Eclipse's JDT.

The problem is that it is very common to pass in an object to a method,
say an IProgressMonitor, and the parameter can be supplied as null. But
the method wants to always assume its not-null, as this simplifies the
code quite a bit.  So the method tests for the null parameter right away
and reassigns it to a NullProgressMonitor instance.  But setting up a
new local variable to hold either the input parameter value or the new
NullProgressMonitor is asking for trouble, as the parameter is still
visible within the scope of the method.  Its possible to have Eclipse
autocomplete the parameter, when really the local needed to be used.

So what I do here is I'm *very* much in the habit of declaring anything
final; parameters, local variables, instance members, etc, and I only
ever declare something to be not final when I really do intend on
modifying it within its scope.  Its unfortunate that final isn't the
default (or cannot be made to be the default), and its also unfortunate
that a variable cannot be declared to be final below a certain point
in a method.  Such is the quality of programming languages today.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>