Documented operations in egit.core.
This change documents the public classes and constructors in the EGit
core plugin, resolving all warnings associated with them.

I also simplified the implemenetation of the CheckpointOperation,
as it was longer and more complex than it really needed to be.  It
also missed the critical condition where GitProjectData can go missing
and get(IProject) returns null; normally we just skip over these and
don't attempt to work with that project in the current operation.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>