Remove unnecessary { and } from statements.
This was just annoying me.  One of the coding standards I work with
requires { and } in all situations, but I'm finding it just makes the
code take up more vertical space (thanks to the } on its own line)
and doesn't really contribute to the readability.

There should not be any logic changes here, I've just whacked the
extra braces.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>