Rename mergeCurrentTree to mergeSubtree.
We are in the process of merging what most people would consider the
current tree; its the set of Trees we were given in the constructor.
The variable mergeCurrentTree is actually controlling whether or not
we need to expend time recursively merging any tree this tree points
at.  Renaming it to mergeSubtree just makes more sense.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>