Revert "Fix merge of subtrees"
This reverts commit ca962feae42f00fb0228d2ef3b64ffa1d9740739.

Adding in !mergeSubtree here causes us to expand the subtree
sometimes even when its unchanged.  For performance reasons
we want to avoid touching trees which are not different, as
this can really reduce the amount of work required on very
big projects.

The real bug that ca96 was trying to fix is that findMember
returns null if the member in question is contained in a
subtree that has no changes in it.  It still needs to be
fixed.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>