Don't aggressively merge subtrees.
There is no reason to merge a subtree recursively just because it is
dirty.  The user of a MergedTree instance may never bother to go into
a subtree and examine its contents, even if it is different.

I'm also renaming computeMerge to matchByName.  What we are really
doing is simply matching the TreeEntry objects by name across the
various input trees; we are not performing any sort of traditional
merge action.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>