Document GitResourceDecorator and make refresh slightly easier to call.
While documenting this class, I decided to make the refresh
method make use of the RCL class's ability to schedule
a decorator update on the UI thread if the caller is not
the UI thread.  This makes it easier for code which may not
be on the UI thread to request a decorator refresh.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>