Relicense org.spearce.jgit under GPL.
After some discussion on #git with Robin Rosenburg about the difficulty
of porting concepts from core "C" based Git to jgit we have decided to
license org.spearce.jgit (the pure Java implementation of Git) under
the GNU Public License, as this is the same license that covers the
core Git code.  By using the same license as the core Git code we can
easily port algorithms written by others while maintaing their wishes
relative to the license associated with them.

The other packages are staying a mixture of LGPL and EPL as the code
there is written by people beyond just Robin and myself, and this code
must interface with Eclipse.

Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>
Signed-off-by: Shawn O. Pearce <spearce@spearce.org>