Cache commits
This uses more memory, but by having a WeakMap we allow the
garbage collector to claim the memory if needed. This speeds
up the history panel enormously, so that having the history
pane connected to the selected resource is not a performance
problem (except the first time the history of a resource in
a projects is shown).

Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>
Signed-off-by: Shawn O. Pearce <spearce@spearce.org>