Add support for structured comparison.
With this code we can now compare revisions at project and folder
levels, rather than only file level. We use the compare api's that
eclipse provide to do the display and detailed compare work, while
we do the high level comparisons at the SHA-1 level.

A possible downside is that the API:s we use are internal, which
could make us version dependent. This code works with both 3.2.x and
3.3M5 versions though.

Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>
Signed-off-by: Shawn O. Pearce <spearce@spearce.org>