Move GitResourceNode to ui plugin
Because I am trying to keep the core plugin UI free, allowing it
to be used in headless workbenches, such as from an Eclipse based
build system, we don't want to depend on the org.eclipse.swt plugin.
Relocating GitResourceNode to the UI plugin lets us continue to avoid
that dependency.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>