Use the Git sort order.
Do not resort trees when reading into memory. This requires
that a number of routines learn to sort tree
entries as if they end with a '/'. To do that a parameter
must be passed. I do this by passing a NUL or '/' byte as
an extra parameter for the comparison routines.

No noticeable performance difference can be noted. The reason
for this patch is to simply implementation of the git index
and make jgit more git-like.

Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>
Signed-off-by: Shawn O. Pearce <spearce@spearce.org>