Recursion and update of all elements regardless of need.
When selecting an element in the history view viewer.updateElement
is called for every element recursively. If the number of
elements is very largs this results in a stack overflow. This
only happens with Eclipse 3.3 M7 and not M6.

Is it thig bug? https://bugs.eclipse.org/bugs/show_bug.cgi?id=149642

Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>
Signed-off-by: Shawn O. Pearce <spearce@spearce.org>