Discard the index file handle after reading it
Now that we store the entire index file in a byte[] and no
longer access it on demand through the WindowedFile handle
there is no reason to maintain the WindowedFile handle or
any cached windows that may be associated with it.  Better
that we discard that file as quickly as possible than to
hang onto it and all of the associated resources, as we
could wind up with the index data in virtual memory twice.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>