Introduce the Git index
This is a partial and non-optimized version of the Git Index. It covers
reading and writing of index with stage 0 entries from the index file plus
reading from trees plus writing trees. Adding and removing entries as well
as refreshing the timestamp and execute bit of the stat information is
supported. JDK1.6 is required to support the execute bit though. Most
of the stat information is ignored as the Java API does not have any
support for things like i-node number, device id or uid/gid information.

It is the same index as native Git uses so it should be enough to use
native Git, Stacked Git and Eclipse in paralell.

Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>