Replace writeRef with RefLock instances
Our jgit RefLock class is meant to implement the same locking policy
that core Git uses, meaning it should be safe to run both jgit and
core Git on the same repository at the same time and not run into lost
updates due to locking problems.

Unfortunately Robin's implementation of writeRef doesn't use that same
policy, which makes it harder to interoperate nicely with core Git.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>