Flush RefLock buffers before releasing
We hold a whole-file lock while we are supposed to be updating the lock
file.  Once we release it, we really shouldn't be writing into the file.
But we are writing to a buffer, which is probably not going to flush to
the file until we close it, or force it to flush.  So we need to flush
before the release, just to be sure our data got to the file before the
lock was removed.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>