Allow RefLock to be used for other files
When we create a symbolic-ref (for example) we want to use the same
locking system as we do for normal refs, as that is what core Git does,
and its just nice to reuse code that is already thought to be stable.

Here I'm just exposing the constructor to allow creation of a RefLock
against any file path, not just one that is a Ref.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>