Ignore unqualified field access warnings
Robin earlier changed this setting to warning, but we have thousands
of such cases in our current code.  Tweaking a warning on doesn't have
a lot of value when it adds thousands of warnings in Eclipse as its
nearly impossible to see newer warnings, or possibly more important
warnings.

Personally I see the unqualified field access to be a feature of any
modern object oriented language.  The implicit "this." (or this->)
saves keystrokes and is one of the major features that an OO system
has over a procedural one.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>