Cleanup synthetic-access warnings in all current locations
Most of these call sites are because we are trying to access a
private member of another class, usually from within an inner
class.  In this case Java compilers actually generate another
method that is public to perform the access, slightly slowing
down the codepath.

In the case of these constructor calls in FileMode we are making
constants, and don't want anyone else to be able to create new
instances of this class.  Keeping the constructor private is the
best way to do this, but I don't want to see these warnings anymore.
@SuppressWarnings("synthetic-access") is a great way to hide these.

RepositoryConfig's Entry inner class is sort of on the critical
path of reading a config file, so I'm making it package level access.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>