Correct "variable may be null" errors
In at least one of these locations detected by Eclipse it was
possible for the variable in question to be null, causing a
possible NullPointerException.  Since NPEs are generally bad we
should avoid it.

A few other locations are easy to correct too, but are just a
minor cleanup of the control structures.  Apparently the Eclipse
data flow analysis doesn't do a great job with do {} while loops,
but works fine with for (;;) {if () break;} which is also a very
common way of writing these types of expressions.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org