Ignore parameter assignments within methods in jgit
I *always* delcare method parameters final unless I want to
be able to alter them within the body of the method.  This
is actually rather rare, but when I want to change a parameter
within the method its usually the clearer way to do implement
the desired logic.  Getting warnings over such cases defeats
the purpose of trying to produce very clear code that has no
bugs within it.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>