Remove warnings about unnecessary semicolons
These warnings are useful, and yet are being drowned out by the
mass of warnings we are getting that are less useful, like not
having documentation in test cases.  *sigh*  This is a perfect
example of how its very easy to shoot yourself in the foot by
making compiler warnings too tight too fast.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>