Rewrite checkpointjob for index
This makes CheckpointJob do something similar with the index that it used to do with the
tree "index". All files managed by the index will be updated including other projects
and files not visible in eclipse, as long as they belong to the same index as the resources
that we given to checkpoint.

There is no more implicit invocation of CheckpointJob.

Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>