Commiting works with multiple projects regardless of repo root
When fixing the multi-repository case, I broke the single-repository one.
Now, it checks to see if the repository is rooted at the project or workspace level.
If you have something other than that, tough luck - I can't even imagine how it
would be possible to handle something like that - or even worse, nested repositories.
At any rate, this should cover about 99% of the cases.

Signed-off-by: Dave Watson <dwatson@mimvista.com>